modificando el acceso a companyPage
